리펙토링 : 전체 페키징 처리
