+++ src/main/java/bowling/view/Input.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.view;
 
 import java.io.InputStream;
 import java.util.Optional;
