+++ src/main/java/bowling/view/Output.java
@@ -1,11 +1,11 @@
-package bowling;
+package bowling.view;
 
 import java.util.stream.IntStream;
 
-import static bowling.BowlingUtils.BEGIN_FORMAT;
-import static bowling.BowlingUtils.ING_FORMAT;
-import static bowling.BowlingUtils.MAX_SCORE;
-import static bowling.BowlingUtils.MIN_SCORE;
+import static bowling.utils.BowlingUtils.BEGIN_FORMAT;
+import static bowling.utils.BowlingUtils.ING_FORMAT;
+import static bowling.utils.BowlingUtils.MAX_SCORE;
+import static bowling.utils.BowlingUtils.MIN_SCORE;
 
 public class Output {
